Refactoring and Cleanup
