+++ simnibs/simulation/__init__.py
@@ -1,7 +1,7 @@
-'''
+"""
 SimNIBS simulation handling
 
-'''
+"""
 
 from .sim_struct import *
 from .fem import *
